+++ src/operations/uninitialised/relu.rs
@@ -38,8 +38,8 @@ impl UninitialisedOperation for Operation {
     fn with_iter_private(
         self,
         _iter: &mut impl Iterator<Item = ElementType>,
-        input_neuron_count: usize,
-    ) -> Result<(Self::Initialised, usize)> {
+        input_neuron_count: u16,
+    ) -> Result<(Self::Initialised, u16)> {
         Ok((
             initialised::relu::Operation {
                 neurons: input_neuron_count,
@@ -49,11 +49,7 @@ impl UninitialisedOperation for Operation {
         ))
     }
 
-    fn with_seed_private(
-        self,
-        _seed: u64,
-        input_neuron_count: usize,
-    ) -> (Self::Initialised, usize) {
+    fn with_seed_private(self, _seed: u64, input_neuron_count: u16) -> (Self::Initialised, u16) {
         (
             initialised::relu::Operation {
                 neurons: input_neuron_count,
